+++ Ranger/App.swift
@@ -16,59 +16,45 @@ class App: TableTrackerDelegate
     
     private var tablesStatusBarItem: TablesStatusBarItem = TablesStatusBarItem()
     private var windowTracker: TableTracker = TableTracker()
-    private var appWindow: NSWindow?
+    private var tableWindowController: TourneyTableWindowController?
     
     
     func start()
     {
-        // First window is status bar item.
-        self.appWindow = NSApplication.shared.windows[1]
-        
         // Subscribe window updates.
         windowTracker.delegate = self
         windowTracker.start()
     }
     
     
-    // MARK: - WindowTracker
+    // MARK: - Track tables
     
     func windowTrackerDidStartTrackingTable(tableWindowInfo: TableWindowInfo)
     {
-        // Prevent drag.
-        appWindow?.isMovable = false
-        
-        // UI.
+        // Status.
         tablesStatusBarItem.indicateTracking(windowTitle: tableWindowInfo.name)
+        
+        // Window.
+        tableWindowController = TourneyTableWindowController.instantiateAndShow(forTableWindowInfo: tableWindowInfo)
     }
     
     func windowTrackerDidUpdateTableWindowInfo(tableWindowInfo: TableWindowInfo)
     {
         // Only if any.
-        guard let appWindow = appWindow
-        else { return }
+        guard let tableWindowController = tableWindowController else { return }
         
-        // Align.
-        appWindow.setFrame(
-            NSRect(
-                x: tableWindowInfo.UIKitBounds.origin.x,
-                y: tableWindowInfo.UIKitBounds.origin.y - appWindow.frame.size.height,
-                width: tableWindowInfo.UIKitBounds.size.width,
-                height: appWindow.frame.size.height
-            ),
-            display: true
-        )
-                 
-         // Put above.
-         appWindow.order(NSWindow.OrderingMode.above, relativeTo: tableWindowInfo.number)
+        // Update.
+        tableWindowController.update(with: tableWindowInfo)
     }
     
     func windowTrackerDidStopTrackingTable(tableWindowInfo: TableWindowInfo)
     {
-        // Enable drag.
-        appWindow?.isMovable = true
-        
         // UI.
         tablesStatusBarItem.stopIndicateTracking()
+        
+        // Close window if any.
+        if let tableWindowController = tableWindowController
+        { tableWindowController.close() }
     }
     
 }